rrr: convert base/strop + base/basetypes to merged-module form
Switch to the merged-pair approach: each .hpp/.cpp pair collapses into
a single .cpp module interface unit; the .hpp is deleted. Class members
declared in a header (global module) and defined in a .cpp module
(rrr.X attachment) hit libc++/libc ODR errors as soon as they pull
common headers from both attachments — the shim approach we tried for
strop only works for free-function-only files. Merging avoids the split
entirely.

Changes:

  src/rrr/base/strop.cpp
    - Module unit (already converted) is unchanged. Removed the
      now-orphan strop.hpp.

  src/rrr/base/basetypes.cpp
    - Rewritten as the rrr.basetypes module: class declarations
      (SparseInt, v32, v64, NoCopy, RefCounted, Counter, Time, Timer,
      Rand) + templates (atomic_*, Enumerator, MergedEnumerator) +
      out-of-line definitions all live in this file.
    - `verify(...)` macro/template from debugging.hpp dropped; replaced
      with inline `if (!cond) std::abort();` (4 call sites). Avoids
      pulling debugging.hpp's std_compat chain into the module.

  src/rrr/base/strop.hpp, src/rrr/base/basetypes.hpp — deleted.

Importing sites updated:
  - src/rrr/rrr.hpp: drop `#include "base/{strop,basetypes}.hpp"`;
    add `import rrr.strop;` + `import rrr.basetypes;` at the bottom
    of the textual chain. libc++ rejects the reverse order
    (`import std; ... #include <vector>` lands the include after the
    module's already-imported std).
  - src/rrr/base/all.hpp: same pattern.
  - src/rrr/base/misc.hpp: replace include with import.
  - src/rrr/base/threading.hpp: replace include with import.
  - src/rrr/rpc/request_queue.hpp: replace include with import.
  - src/rrr/reactor/quorum_event.h: replace include with import.
  - src/rrr/reactor/fiber.h: replace include with import.
  - src/rrr/rpc/idempotency.hpp: drop `class Counter;` forward-decl
    (now in module attachment, can't be re-declared in global module);
    add `import rrr.basetypes;` instead.

Transitive-include repairs surfaced by removing strop.hpp from the
chain:
  - src/rrr/misc/dball.hpp: add `#include <rusty/function.hpp>`
    (was relying on strop.hpp → std_compat → rusty for Function).
  - src/rrr/base/misc.hpp: same.

CMakeLists.txt: basetypes.cpp added to RRR_MODULE_SRC.

Measurement (rrr + rpcbench, -j32, clean):
  baseline:       16.86 s, 28.5 MB PCM (std.pcm + std.compat.pcm)
  strop:          20.65 s, 53.0 MB PCM (+rrr.strop, +24.5 MB)
  strop+basetypes: 20.66 s, 77.6 MB PCM (+rrr.basetypes, +24.6 MB)

Wallclock flat vs strop alone — basetypes BMI builds in parallel.
+22% vs baseline; abort threshold is 3x (~50 s). Plan doc updated.
