+++ src/rrr/CMakeLists.txt
@@ -17,6 +17,7 @@ file(GLOB RRR_BENCH_SRC
 # declaration shim for non-importing consumers; the linker resolves
 # their calls to the module-emitted .o.
 set(RRR_MODULE_SRC
+    ${CMAKE_CURRENT_SOURCE_DIR}/base/basetypes.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/base/strop.cpp
 )
 list(REMOVE_ITEM RRR_BENCH_SRC ${RRR_MODULE_SRC})
@@ -47,7 +48,7 @@ if(NOT TARGET rrr)
         ${CMAKE_CURRENT_SOURCE_DIR}/reactor/reactor.cc
         ${CMAKE_CURRENT_SOURCE_DIR}/base/logging.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/base/misc.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/base/basetypes.cpp
+        # base/basetypes.cpp omitted — module unit (see RRR_MODULE_SRC).
         ${CMAKE_CURRENT_SOURCE_DIR}/base/debugging.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/base/threading.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/misc/alock.cpp
