+++ src/rrr/base/all.hpp
@@ -3,9 +3,12 @@
 //#include "unittest.hpp"
 
 
-#include "basetypes.hpp"
 #include "debugging.hpp"
 #include "logging.hpp"
 #include "misc.hpp"
-#include "strop.hpp"
 #include "threading.hpp"
+
+// Imports go after textual #includes (libc++ rejects `import std; ...
+// #include <vector>` order).
+import rrr.basetypes;
+import rrr.strop;
